feat(authorization): jwt claims and authorization

+++ resources/item.py
@@ -2,7 +2,7 @@
 """Items API resource. """
 from flask.views import MethodView
 from flask_smorest import Blueprint, abort
-from flask_jwt_extended import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt
 from sqlalchemy.exc import SQLAlchemyError
 
 from db import db
@@ -42,6 +42,9 @@ class Item(MethodView):
     @jwt_required()
     def delete(self, item_id):
         """Deletes an item by its id."""
+        jwt = get_jwt()
+        if not jwt.get("is_admin"):
+            abort(403, message="Admin privilege required")
         item = ItemModel.query.get_or_404(item_id)
         db.session.delete(item)
         db.session.commit()
